finalized products without mongo
